# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OncoSafeRx API (New Clinical/Genomic/AI Endpoints)
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Treatment:
      type: object
      required: [regimen]
      properties:
        regimen: { type: string }
        type: { type: string, enum: [chemotherapy, immunotherapy, targeted, combination, radiation, surgery] }
        agents: { type: array, items: { type: string } }
        dosing: { type: object }
        schedule: { type: string }
        duration: { type: number }
        evidence: { type: string, enum: [A, B, C, D] }
        approvalStatus: { type: string }
    PatientData:
      type: object
      required: [age, sex, cancerType, cancerStage, performanceStatus]
      properties:
        patientId: { type: string }
        age: { type: integer }
        sex: { type: string, enum: [male, female, other] }
        cancerType: { type: string }
        cancerStage: { type: string }
        histology: { type: string }
        performanceStatus: { type: integer, minimum: 0, maximum: 4 }
        comorbidities: { type: array, items: { type: string } }
        biomarkers: { type: object, additionalProperties: true }
        genomics: { type: object, additionalProperties: true }
        priorTreatments:
          type: array
          items:
            type: object
            properties:
              regimen: { type: string }
              response: { type: string }
              duration: { type: number }
              endReason: { type: string }
        laboratoryValues: { type: object, additionalProperties: true }
        kidneyFunction: { type: number }
        liverFunction: { type: object, additionalProperties: true }
        cardiacFunction: { type: object, additionalProperties: true }
        metastaticSites: { type: array, items: { type: string } }
paths:
...
  /ai/treatment/predict-treatment:
    post:
      tags: [AI]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientData, availableTreatments]
              properties:
                patientData: { $ref: '#/components/schemas/PatientData' }
                availableTreatments:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/Treatment' }
                predictionOptions: { type: object }
      responses:
        '200': { description: AI treatment prediction }
  /ai/treatment/predict-response:
    post:
      tags: [AI]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientData, treatment]
              properties:
                patientData: { $ref: '#/components/schemas/PatientData' }
                treatment: { $ref: '#/components/schemas/Treatment' }
                responseMetrics:
                  type: array
                  items: { type: string, enum: [ORR, DCR, PFS, OS, duration] }
      responses:
        '200': { description: Response prediction }
  /ai/treatment/predict-survival:
    post:
      tags: [AI]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientData, treatments]
              properties:
                patientData: { $ref: '#/components/schemas/PatientData' }
                treatments:
                  type: array
                  items: { $ref: '#/components/schemas/Treatment' }
                survivalMetrics:
                  type: array
                  items: { type: string, enum: [OS, PFS, DFS, RFS] }
                timePoints:
                  type: array
                  items: { type: integer }
      responses:
        '200': { description: Survival prediction }
  /ai/treatment/predict-toxicity:
    post:
      tags: [AI]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientData, treatment]
              properties:
                patientData: { $ref: '#/components/schemas/PatientData' }
                treatment: { $ref: '#/components/schemas/Treatment' }
                toxicityTypes:
                  type: array
                  items:
                    type: string
                    enum: [hematologic, gastrointestinal, neurologic, cardiac, hepatic, renal, dermatologic]
                severityLevels:
                  type: array
                  items:
                    type: string
                    enum: [grade1, grade2, grade3, grade4, grade5]
      responses:
        '200': { description: Toxicity prediction }
  /ai/treatment/compare-treatments:
    post:
      tags: [AI]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientData, treatments]
              properties:
                patientData: { $ref: '#/components/schemas/PatientData' }
                treatments:
                  type: array
                  minItems: 2
                  items: { $ref: '#/components/schemas/Treatment' }
                comparisonCriteria: { type: object }
                includeEconomicAnalysis: { type: boolean }
      responses:
        '200': { description: Comparison and ranking }
  /ai/treatment/explain-prediction:
    post:
      tags: [AI]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [predictionId]
              properties:
                predictionId: { type: string }
                explanationType: { type: string, enum: [detailed, summary, patient_friendly] }
                includeVisualization: { type: boolean }
      responses:
        '200': { description: Explanation payload }